+++ libargos/inspector/pgplugins/imageplot2d.py
@@ -22,28 +22,17 @@ from __future__ import division, print_function
 import logging
 import pyqtgraph as pg
 
-logger = logging.getLogger(__name__)
-
-USE_SIMPLE_PLOT = False
-
-if USE_SIMPLE_PLOT:
-    # An experimental simplification of PlotItem. Not included in the distribution
-    logger.warn("Using SimplePlotItem as PlotItem")
-    from pyqtgraph.graphicsItems.PlotItem.simpleplotitem import SimplePlotItem
-else:
-    from pyqtgraph.graphicsItems.PlotItem import PlotItem as SimplePlotItem
-
 from libargos.info import DEBUGGING
 from libargos.config.groupcti import MainGroupCti
 from libargos.config.boolcti import BoolCti
 from libargos.config.choicecti import ChoiceCti
 from libargos.inspector.abstract import AbstractInspector
-from libargos.inspector.pgplugins.pgctis import PgPlotItemCti, PgAxisLabelCti
+from libargos.inspector.pgplugins.pgctis import PgPlotItemCti, PgAxisLabelCti, PgAxisFlipCti
+from libargos.inspector.pgplugins.pgctis import PgAxisShowCti, X_AXIS, Y_AXIS
+from libargos.inspector.pgplugins.pgplotitem import ArgosPgPlotItem
 from libargos.utils.cls import array_has_real_numbers, check_class
 
 logger = logging.getLogger(__name__)
-
-
 
 
 class PgImagePlot2dCti(MainGroupCti):
@@ -68,15 +57,18 @@ class PgImagePlot2dCti(MainGroupCti):
         self.plotItemCti = self.insertChild(PgPlotItemCti(plotItem))
 
         xAxisCti = self.plotItemCti.xAxisCti
+        #xAxisCti.insertChild(PgAxisShowCti(plotItem, 'bottom')) # disabled, seems broken
         xAxisCti.insertChild(PgAxisLabelCti(plotItem, 'bottom', self.pgImagePlot2d.collector,
-                                            configValues=["{x-dim}"]))
+            defaultData=1, configValues=[PgAxisLabelCti.NO_LABEL, "{x-dim}"]))
+        xAxisCti.insertChild(PgAxisFlipCti(plotItem, X_AXIS))
 
         yAxisCti = self.plotItemCti.yAxisCti
+        #yAxisCti.insertChild(PgAxisShowCti(plotItem, 'left'))  # disabled, seems broken
         yAxisCti.insertChild(PgAxisLabelCti(plotItem, 'left', self.pgImagePlot2d.collector,
-                                            configValues=["{y-dim}"]))
+            defaultData=1, configValues=[PgAxisLabelCti.NO_LABEL, "{y-dim}"]))
+        yAxisCti.insertChild(PgAxisFlipCti(plotItem, Y_AXIS))
 
         self.insertChild(BoolCti('auto levels', True))
-
 
 
 
@@ -90,8 +82,8 @@ class PgImagePlot2d(AbstractInspector):
         super(PgImagePlot2d, self).__init__(collector, parent=parent)
 
         self.viewBox = pg.ViewBox(border=pg.mkPen("#000000", width=1))
-        self.plotItem = SimplePlotItem(name='1d_line_plot_#{}'.format(self.windowNumber),
-                                       enableMenu=False, viewBox=self.viewBox)
+        self.plotItem = ArgosPgPlotItem(name='2d_image_plot_#{}'.format(self.windowNumber),
+                                        enableMenu=False, viewBox=self.viewBox)
         self.viewBox.setParent(self.plotItem)
 
         self.imageItem = pg.ImageItem()
@@ -101,7 +93,7 @@ class PgImagePlot2d(AbstractInspector):
         self.histLutItem.setImageItem(self.imageItem)
 
         self.graphicsLayoutWidget = pg.GraphicsLayoutWidget()
-        self.titleLabel = self.graphicsLayoutWidget.addLabel('My label', 0, 0, colspan=2)
+        self.titleLabel = self.graphicsLayoutWidget.addLabel('<Title label>', 0, 0, colspan=2)
         self.graphicsLayoutWidget.addItem(self.plotItem, 1, 0)
         self.graphicsLayoutWidget.addItem(self.histLutItem, 1, 1)