+++ libargos/inspector/pgplugins/pgctis.py
@@ -21,19 +21,24 @@ from __future__ import division, print_function
 
 import logging
 import numpy as np
+import pyqtgraph as pg
 
 from libargos.config.groupcti import GroupCti
 from libargos.config.boolcti import BoolCti
 from libargos.config.choicecti import ChoiceCti
-from libargos.config.floatcti import SnFloatCti
+from libargos.config.floatcti import SnFloatCti, FloatCti
 from libargos.config.untypedcti import UntypedCti
 from libargos.info import DEBUGGING
 from libargos.utils.cls import check_class
 
 logger = logging.getLogger(__name__)
 
+X_AXIS = pg.ViewBox.XAxis
+Y_AXIS = pg.ViewBox.YAxis
+VALID_AXIS_POSITIONS =  ('left', 'right', 'bottom', 'top')
 
-class ViewBoxDebugCti(GroupCti): # TODO: somehow don't save the state for this class
+
+class ViewBoxDebugCti(GroupCti):
     """ Read-only config tree for inspecting a PyQtGraph ViewBox
     """
     def __init__(self, nodeName, viewBox, expanded=False):
@@ -91,97 +96,11 @@ class ViewBoxDebugCti(GroupCti): # TODO: somehow don't save the state for this c
                     limitChildItem.data = limitValue
 
 
-#
-# class PgAxisRangeCti(GroupCti):
-#     """ Configuration tree item is linked to the axis range.
-#         Can toggle the axis' auto-range property on/off by means of checkbox
-#     """
-#     def __init__(self, viewBox, axisNumber, nodeName='range'):
-#         """ Constructor.
-#             The monitored axis is specified by viewBox and axisNumber (0 for x-axis, 1 for y-axis)
-#         """
-#         super(PgAxisRangeCti, self).__init__(nodeName)
-#         assert axisNumber in (0, 1), "axisNumber must be 0 or 1"
-#         self.viewBox = viewBox
-#         self.axisNumber = axisNumber
-#
-#         self.rangeMinItem = self.insertChild(SnFloatCti('min', 0.0))
-#         self.rangeMaxItem = self.insertChild(SnFloatCti('max', 1.0))
-#         self.autoRangeItem = self.insertChild(BoolCti("auto-range", True))
-#
-#
-#     def _refreshNodeFromTarget(self):
-#         """ Refreshes the config values from the axes' state
-#         """
-#         # Set the precision from by looking how many decimals are neede to show the difference
-#         # between the minimum and maximum, given the maximum. E.g. if min = 0.04 and max = 0.07,
-#         # we would only need zero decimals behind the point as we can write the range as
-#         # [4e-2, 7e-2]. However if min = 1.04 and max = 1.07, we need 2 decimals behind the point.
-#         # So while the range is the same size, we need more decimals because we are not zoomed in
-#         # around zero.
-#         rangeMin, rangeMax = self.viewBox.state['viewRange'][self.axisNumber]
-#         maxOrder = int(np.log10(np.abs(max(rangeMax, rangeMin))))
-#         diffOrder = int(np.log10(np.abs(rangeMax - rangeMin)))
-#
-#         extraDigits = 2 # add some extra digits to make each pan/zoom action show a difference.
-#         precision = min(25, max(extraDigits + 1, abs(maxOrder - diffOrder) + extraDigits)) # clip
-#         #logger.debug("maxOrder: {}, diffOrder: {}, precision: {}"
-#         #             .format(maxOrder, diffOrder, precision))
-#
-#         self.rangeMinItem.precision = precision
-#         self.rangeMaxItem.precision = precision
-#
-#         self.rangeMinItem.data, self.rangeMaxItem.data = rangeMin, rangeMax
-#         autoRangeEnabled = bool(self.viewBox.autoRangeEnabled()[self.axisNumber])
-#
-#         #logger.debug("PgAxisRangeCti.refresh: axis {}, {}".format(self.axisNumber, autoRangeEnabled))
-#         self.rangeMinItem.enabled = not autoRangeEnabled
-#         self.rangeMaxItem.enabled = not autoRangeEnabled
-#         #self.autoRangeItem.data = autoRangeEnabled
-#
-#         self.model.emitDataChanged(self)
-#
-#
-#     def getRange(self):
-#         """Returns a tuple with the min and max range
-#         """
-#         return (self.rangeMinItem.data, self.rangeMaxItem.data)
-#
-#
-#     def _setAutoRange(self):
-#         """ Sets the auto range
-#         """
-#         autoRange = self.autoRangeItem.configValue
-#         logger.debug("enableAutoRange: axis {}, {}".format(self.axisNumber, autoRange))
-#         #assert False, "stopped"
-#
-#         if autoRange:
-#             self.viewBox.enableAutoRange(self.axisNumber, autoRange)
-#         else:
-#             if self.axisNumber == self.viewBox.XAxis:
-#                 xRange, yRange = self.getRange(), None
-#             else:
-#                 xRange, yRange = None, self.getRange()
-#
-#             logger.debug("setRange: xRange={}, yRange={}".format(xRange, yRange))
-#
-#             # viewBox.setRange doesn't accept an axis number :-(
-#             self.viewBox.setRange(xRange = xRange, yRange=yRange,
-#                                   padding=0, update=False, disableAutoRange=True)
-#
-#
-#     def _updateTargetFromNode(self):
-#         """ Applies the configuration to the target axis it monitors.
-#         """
-#         self._setAutoRange()
-#
-#
-
 
 class PgAxisLabelCti(ChoiceCti):
-    """ Configuration tree item is linked to the axis label.
+    """ Configuration tree item that is linked to the axis label.
     """
-    VALID_AXIS_POSITIONS =  ('left', 'right', 'bottom', 'top')
+    NO_LABEL = '-- none --'
 
     def __init__(self, plotItem, axisPosition, collector,
                  nodeName='label', defaultData=0, configValues=None):
@@ -195,8 +114,8 @@ class PgAxisLabelCti(ChoiceCti):
         """
         super(PgAxisLabelCti, self).__init__(nodeName, editable=True,
                                              defaultData=defaultData, configValues=configValues)
-        assert axisPosition in self.VALID_AXIS_POSITIONS, \
-            "axisPosition must be in {}".format(self.VALID_AXIS_POSITIONS)
+        assert axisPosition in VALID_AXIS_POSITIONS,\
+            "axisPosition must be in {}".format(VALID_AXIS_POSITIONS)
         self.collector = collector
         self.plotItem = plotItem
         self.axisPosition = axisPosition
@@ -204,9 +123,84 @@ class PgAxisLabelCti(ChoiceCti):
 
     def _updateTargetFromNode(self):
         """ Applies the configuration to the target axis it monitors.
+            The axis label will be set to the configValue. If the configValue equals
+             PgAxisLabelCti.NO_LABEL, the label will be hidden.
         """
         rtiInfo = self.collector.getRtiInfo()
         self.plotItem.setLabel(self.axisPosition, self.configValue.format(**rtiInfo))
+        self.plotItem.showLabel(self.axisPosition, self.configValue != self.NO_LABEL)
+
+
+
+class PgAxisShowCti(BoolCti):
+    """ BoolCti that toggles axisPosition showing/hiding an axis.
+    """
+    def __init__(self, plotItem, axisPosition, nodeName='show', defaultData=True):
+        """ Constructor.
+            The target axis is specified by plotItem and axisPosition.
+            axisPosition must be one of: 'left', 'right', 'bottom', 'top'
+
+            NOTE: the PyQtGraph showAxis seems not to work. TODO: investigate.
+        """
+        super(PgAxisShowCti, self).__init__(nodeName, defaultData=defaultData)
+        assert axisPosition in VALID_AXIS_POSITIONS,\
+            "axisPosition must be in {}".format(VALID_AXIS_POSITIONS)
+
+        self.plotItem = plotItem
+        self.axisPosition = axisPosition
+
+
+    def _updateTargetFromNode(self):
+        """ Applies the configuration to its target axis
+        """
+        logger.debug("showAxis: {}, {}".format(self.axisPosition, self.configValue))
+        self.plotItem.showAxis(self.axisPosition, show=self.configValue) # Seems not to work
+
+
+class PgAxisLogModeCti(BoolCti):
+    """ BoolCti that toggles an axis between logarithmic vs linear mode.
+    """
+    def __init__(self, plotItem, axisNumber, nodeName='logarithmic', defaultData=False):
+        """ Constructor.
+            The target axis is specified by plotItem and axisNumber (0 for x-axis, 1 for y-axis)
+        """
+        super(PgAxisLogModeCti, self).__init__(nodeName, defaultData=defaultData)
+        self.plotItem = plotItem
+        self.axisNumber = axisNumber
+
+
+    def _updateTargetFromNode(self):
+        """ Applies the configuration to its target axis
+        """
+        if self.axisNumber == X_AXIS:
+            xMode, yMode = self.configValue, None
+        else:
+            xMode, yMode = None, self.configValue
+
+        self.plotItem.setLogMode(x=xMode, y=yMode)
+
+
+class PgAxisFlipCti(BoolCti):
+    """ BoolCti that flips an axis when True.
+    """
+    def __init__(self, plotItem, axisNumber, nodeName='flipped', defaultData=False):
+        """ Constructor.
+            The target axis is specified by plotItem and axisNumber (0 for x-axis, 1 for y-axis)
+        """
+        super(PgAxisFlipCti, self).__init__(nodeName, defaultData=defaultData)
+        assert axisNumber in (X_AXIS, Y_AXIS), "axisNumber must be 0 or 1"
+        self.plotItem = plotItem # TODO: viewbox parameter instead of plotItem?
+        self.viewBox = plotItem.getViewBox()
+        self.axisNumber = axisNumber
+
+
+    def _updateTargetFromNode(self):
+        """ Applies the configuration to its target axis
+        """
+        if self.axisNumber == X_AXIS:
+            self.viewBox.invertX(self.configValue)
+        else:
+            self.viewBox.invertY(self.configValue)
 
 
 
@@ -216,11 +210,11 @@ class PgAxisRangeCti(GroupCti):
     """
     def __init__(self, plotItem, axisNumber, nodeName='range'):
         """ Constructor.
-            The monitored axis is specified by viewBox and axisNumber (0 for x-axis, 1 for y-axis)
+            The target axis is specified by plotItem and axisNumber (0 for x-axis, 1 for y-axis)
         """
         super(PgAxisRangeCti, self).__init__(nodeName)
-        assert axisNumber in (0, 1), "axisNumber must be 0 or 1"
-        self.plotItem = plotItem
+        assert axisNumber in (X_AXIS, Y_AXIS), "axisNumber must be 0 or 1"
+        self.plotItem = plotItem # TODO: viewbox parameter instead of plotItem?
         self.viewBox = plotItem.getViewBox()
         self.axisNumber = axisNumber
 
@@ -229,7 +223,6 @@ class PgAxisRangeCti(GroupCti):
         self.autoRangeItem = self.insertChild(BoolCti("auto-range", True))
 
         # Connect signals
-        self.plotItem.autoBtn.clicked.connect(self._setAutoRangeOn)
         self.viewBox.sigRangeChangedManually.connect(self._setAutoRangeOff)
 
 
@@ -237,7 +230,6 @@ class PgAxisRangeCti(GroupCti):
         """ Disconnects signals.
             Is called by self.finalize when the cti is deleted.
         """
-        self.plotItem.autoBtn.clicked.disconnect(self._setAutoRangeOn)
         self.viewBox.sigRangeChangedManually.disconnect(self._setAutoRangeOff)
 
 
@@ -249,7 +241,7 @@ class PgAxisRangeCti(GroupCti):
             The *args and **kwargs arguments are ignored but make it possible to use this as a slot
             for signals with arguments.
         """
-        logger.debug("  {}._refreshNodeFromTarget: {} {}".format(self.nodePath, args, kwargs))
+        #logger.debug("  {}._refreshNodeFromTarget: {} {}".format(self.nodePath, args, kwargs))
         self._refreshAutoRange()
         self._refreshMinMax(self.viewBox, self.viewBox.state['viewRange'])
 
@@ -269,7 +261,7 @@ class PgAxisRangeCti(GroupCti):
             "sanity check failed: {} != {}".format(viewBox, self.viewBox)
 
         rangeMin, rangeMax = ranges[self.axisNumber]
-        logger.debug("    _refreshMinMax: {}, {}".format(rangeMin, rangeMax))
+        #logger.debug("    _refreshMinMax: {}, {}".format(rangeMin, rangeMax))
 
         maxOrder = int(np.log10(np.abs(max(rangeMax, rangeMin))))
         diffOrder = int(np.log10(np.abs(rangeMax - rangeMin)))
@@ -292,20 +284,16 @@ class PgAxisRangeCti(GroupCti):
         """ The min and max config items will be disabled if auto range is on.
         """
         enabled = self.autoRangeItem.configValue
-        logger.debug("    _refreshAutoRange: enabled={}".format(enabled))
+        #logger.debug("    _refreshAutoRange: enabled={}".format(enabled))
         self.rangeMinItem.enabled = not enabled
         self.rangeMaxItem.enabled = not enabled
         self.model.emitDataChanged(self)
 
 
     def _setAutoRangeOn(self):
-        """ Turns on the auto range checkbox.
-            Calls _updateTargetFromNode to calculate the new range.
+        """ Not implemented for single axis. See PgPlotItemCti._setAutorangeOn
         """
-        self.autoRangeItem.data = True
-        self.model.itemChanged.emit(self) # TODO: move to PgPlotItemCti
-        # TODO: reset should go to minimum target
-        self._refreshNodeFromTarget()
+        assert False, "Not implemented for single axis. See PgPlotItemCti._setAutorangeOn"
 
 
     def _setAutoRangeOff(self):
@@ -313,7 +301,6 @@ class PgAxisRangeCti(GroupCti):
             Calls _refreshNodeFromTarget, not _updateTargetFromNode because setting auto range off
             does not require a redraw of the target.
         """
-        model = self.autoRangeItem.model
         self.autoRangeItem.data = False
         self._refreshNodeFromTarget()
 
@@ -328,13 +315,13 @@ class PgAxisRangeCti(GroupCti):
         """ Applies the configuration to the target axis it monitors.
         """
         autoRange = self.autoRangeItem.configValue
-        logger.debug("enableAutoRange: axis {}, {}".format(self.axisNumber, autoRange))
+        #logger.debug("enableAutoRange: axis {}, {}".format(self.axisNumber, autoRange))
 
         if autoRange:
             padding = None # PyQtGraph default: between 0.02 and 0.1 dep. on the size of the ViewBox
             rect = self.viewBox.childrenBoundingRect() # taken from viewBox.autoRange()
             if rect is not None:
-                if self.axisNumber == self.viewBox.XAxis:
+                if self.axisNumber == X_AXIS:
                     xRange, yRange = (rect.left(), rect.right()), None
                 else:
                     xRange, yRange = None, (rect.bottom(), rect.top())
@@ -343,13 +330,13 @@ class PgAxisRangeCti(GroupCti):
                 return
         else:
             padding = 0
-            if self.axisNumber == self.viewBox.XAxis:
+            if self.axisNumber == X_AXIS:
                 xRange, yRange = self.getRange(), None
             else:
                 xRange, yRange = None, self.getRange()
 
         # viewBox.setRange doesn't accept an axis number :-(
-        logger.debug("setRange: xRange={}, yRange={}, padding={}".format(xRange, yRange, padding))
+        #logger.debug("setRange: xRange={}, yRange={}, padding={}".format(xRange, yRange, padding))
         self.viewBox.setRange(xRange = xRange, yRange=yRange, padding=padding,
                               update=False, disableAutoRange=True)
 
@@ -360,19 +347,26 @@ class PgAxisCti(GroupCti):
 
         This CTI is intended to be used as a child of a PgPlotItemCti.
     """
-    def __init__(self, nodeName, plotItem, axisNumber):
-        """ Constructor
+    def __init__(self, plotItem, axisNumber, nodeName):
+        """ Constructor. A PgAxisRangeCti will be inserted by default.
+            The target axis is specified by plotItem and axisNumber (0 for x-axis, 1 for y-axis)
         """
         super(PgAxisCti, self).__init__(nodeName)
 
-        assert axisNumber in (0, 1), "axisNumber must be 0 or 1"
+        assert axisNumber in (X_AXIS, Y_AXIS), "axisNumber must be 0 or 1"
         self.plotItem = plotItem
         self.axisNumber = axisNumber
 
-        #self.insertChild(BoolCti("show", True)) # TODO:
-        #self.insertChild(BoolCti('logarithmic', False))
+        self.rangeItem = self.insertChild(PgAxisRangeCti(self.plotItem, self.axisNumber),
+                                          position=0)
 
-        self.rangeItem = self.insertChild(PgAxisRangeCti(self.plotItem, self.axisNumber))
+
+    def insertChild(self, childItem, position=-1):
+        """ Inserts a childItem.
+            The default position is -1, meaning that it will be inserted at the penultimate
+            position, before the PgAxisRangeCti.
+        """
+        return super(PgAxisCti, self).insertChild(childItem, position=position)
 
 
 
@@ -383,24 +377,25 @@ class PgPlotItemCti(GroupCti):
         """ Constructor
         """
         super(PgPlotItemCti, self).__init__(nodeName)
-
-        #check_class(plotItem, (PlotItem, SimplePlotItem) # TODO
         assert plotItem, "target plotItem is undefined"
 
-        self.aspectItem = self.insertChild(BoolCti("lock aspect ratio", False))
+        # x/y. How do we disable this?
+        self.aspectLockedCti = self.insertChild(BoolCti("lock aspect ratio", False, expanded=False))
+        self.aspectRatioCti = self.aspectLockedCti.insertChild(FloatCti("ratio", 1.0, minValue=0.0))
+
         self.plotItem = plotItem
         viewBox = plotItem.getViewBox()
 
         # Keeping references to the axes CTIs because they need to be updated quickly when
         # the range changes; getting by path may be slow.)
         if xAxisCti is None:
-            self.xAxisCti = self.insertChild(PgAxisCti('x-axis', plotItem, 0))
+            self.xAxisCti = self.insertChild(PgAxisCti(plotItem, X_AXIS, 'x-axis'))
         else:
             check_class(xAxisCti, PgAxisCti)
             self.xAxisCti = self.insertChild(xAxisCti)
 
         if yAxisCti is None:
-            self.yAxisCti = self.insertChild(PgAxisCti('y-axis', plotItem, 1))
+            self.yAxisCti = self.insertChild(PgAxisCti(plotItem, Y_AXIS, 'y-axis'))
         else:
             check_class(yAxisCti, PgAxisCti)
             self.yAxisCti = self.insertChild(yAxisCti)
@@ -411,43 +406,31 @@ class PgPlotItemCti(GroupCti):
             self.stateItem = None
 
         # Connect signals
-        if DEBUGGING:
-            viewBox = self.plotItem.getViewBox()
-            viewBox.sigStateChanged.connect(self.vbStateChanged)
-            viewBox.sigRangeChanged.connect(self.vbRangeChanged)
-            viewBox.sigXRangeChanged.connect(self.vbXRangeChanged)
-            viewBox.sigYRangeChanged.connect(self.vbYRangeChanged)
-            viewBox.sigRangeChangedManually.connect(self.vbRangeChangedManually)
+        self.plotItem.autoBtn.clicked.connect(self._setAutoRangeOn)
+
 
     def _closeResources(self):
         """ Disconnects signals.
             Is called by self.finalize when the cti is deleted.
         """
-        if DEBUGGING:
-            viewBox = self.plotItem.getViewBox()
-            viewBox.sigStateChanged.disconnect(self.vbStateChanged)
-            viewBox.sigRangeChanged.disconnect(self.vbRangeChanged)
-            viewBox.sigXRangeChanged.disconnect(self.vbXRangeChanged)
-            viewBox.sigYRangeChanged.disconnect(self.vbYRangeChanged)
-            viewBox.sigRangeChangedManually.disconnect(self.vbRangeChangedManually)
+        self.plotItem.autoBtn.clicked.disconnect(self._setAutoRangeOn)
 
 
-    def vbStateChanged(self, *args, **kwargs):
-        #logger.debug("viewBox state changed: {} {}".format(args, kwargs))
-        pass
+    def _updateTargetFromNode(self):
+        """ Applies the configuration to the target axis it monitors.
+        """
+        self.plotItem.setAspectLocked(lock=self.aspectLockedCti.configValue,
+                                      ratio=self.aspectRatioCti.configValue)
 
-    def vbRangeChanged(self, *args, **kwargs):
-        #logger.debug("viewBox range changed: {} {}".format(args, kwargs))
-        pass
 
-    def vbXRangeChanged(self, *args, **kwargs):
-        #logger.debug("viewBox X-range changed: {} {}".format(args, kwargs))
-        pass
+    def _setAutoRangeOn(self):
+        """ Turns on the auto range checkbox.
+            Calls _updateTargetFromNode to calculate the new range.
+        """
+        for axisCti in [self.xAxisCti, self.yAxisCti]:
+            for childCti in axisCti.childItems:
+                if isinstance(childCti, PgAxisRangeCti):
+                    childCti.autoRangeItem.data = True
+        self.model.itemChanged.emit(self)
+        #self._refreshNodeFromTarget()
 
-    def vbYRangeChanged(self, *args, **kwargs):
-        #logger.debug("viewBox Y-range changed: {} {}".format(args, kwargs))
-        pass
-
-    def vbRangeChangedManually(self, *args, **kwargs):
-        #logger.debug("viewBox range changed manually: {} {}".format(args, kwargs))
-        pass