+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+Current development branch
+
+
+* Range config tree items (CTI) have scientific notation
+* Axis CTIs can have flip and log-mode.
+
 0.2.0-devel (2016-03-29)
 ---------------------